Fix a potential integer overflow...
...when allocating capacity from a connection level window update to
individual HTTP/2 streams.
Based on #996 by Mike Tingey Jr.

+++ java/org/apache/coyote/http2/Http2UpgradeHandler.java
@@ -1186,7 +1186,8 @@ class Http2UpgradeHandler extends AbstractStream implements InternalHttpUpgradeH
                             break;
                         }
 
-                        int share = (int) (s.getConnectionAllocationRequested() * remaining /
+                        // Cast to long during the intermediate calculation to avoid integer overflow in multiplication
+                        int share = (int) ((long) s.getConnectionAllocationRequested() * remaining /
                                 requestedAllocationForIncrementalStreams);
                         if (share == 0) {
                             share = 1;
